backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.1
markdown==3.5.2 